Extract ContractCallSection
